+++ lib/src/apollovm_code_generator.dart
@@ -4,7 +4,6 @@
 
 import 'apollovm_code_storage.dart';
 import 'apollovm_generator.dart';
-import 'apollovm_utils.dart';
 import 'ast/apollovm_ast_base.dart';
 import 'ast/apollovm_ast_expression.dart';
 import 'ast/apollovm_ast_statement.dart';
@@ -1643,19 +1642,9 @@ abstract class ApolloCodeGenerator
     if (headIndented) out.write(indent);
 
     var v = value.value;
-    var s = v.toString();
+    var s = ASTTypeDouble.doubleToString(v);
 
-    if (v == 0.0) {
-      out.write('0.0');
-    } else if (s.contains('e') || s.contains('E')) {
-      // scientific notation → force decimal format
-      final digits = v.fractionDigitsFromScientificNotation().clamp(0, 20);
-      out.write(v.toStringAsFixed(digits));
-    } else if (!s.contains('.')) {
-      out.write('$s.0');
-    } else {
-      out.write(s);
-    }
+    out.write(s);
 
     return out;
   }